// src/lib/api/model/inboxConversationDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Lean Commerce API
 * The AI Sales Agent CRM API Documentation
 * OpenAPI spec version: 1.0
 */
import type { InboxConversationDtoLead } from "./inboxConversationDtoLead";
import type { InboxMessageDto } from "./inboxMessageDto";

export interface InboxConversationDto {
  id: string;
  organizationId: string;
  /** @nullable */
  leadId: string | null;
  /** @nullable */
  channelId: string | null;
  status: string;
  aiPaused: boolean;
  stateVersion: number;
  /** @nullable */
  assignedAgentId: string | null;
  /** @nullable */
  externalContactId: string | null;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  lead: InboxConversationDtoLead;
  messages: InboxMessageDto[];
}